+++ hooks/pages/index.ts
@@ -1,3 +1,5 @@
 export { useRecipientDetailPage } from './useRecipientDetailPage';
 
 export { useRecipientsListPage } from './useRecipientsListPage';
+
+export { useNewRecipientForm } from './useNewRecipientForm';
